Move `User` from httpClient.ts to user.ts
