+++ app/src/lib/backend/httpClient.ts
@@ -1,4 +1,5 @@
 import { invoke } from './ipc';
+import type { User } from '$lib/stores/user';
 import { PUBLIC_API_BASE_URL } from '$env/static/public';
 
 export type Feedback = {
@@ -15,23 +16,6 @@ export type LoginToken = {
 	expires: string;
 	url: string;
 };
-
-export class User {
-	id!: number;
-	name: string | undefined;
-	given_name: string | undefined;
-	family_name: string | undefined;
-	email!: string;
-	picture!: string;
-	locale!: string;
-	created_at!: string;
-	updated_at!: string;
-	access_token!: string;
-	role: string | undefined;
-	supporter!: boolean;
-	github_access_token: string | undefined;
-	github_username: string | undefined;
-}
 
 export type Project = {
 	name: string;
