+++ app/src/lib/components/BranchCard.svelte
@@ -10,8 +10,6 @@
 	import { AIService } from '$lib/ai/service';
 	import laneNewSvg from '$lib/assets/empty-state/lane-new.svg?raw';
 	import noChangesSvg from '$lib/assets/empty-state/lane-no-changes.svg?raw';
-	import { User } from '$lib/backend/httpClient';
-	import { Project } from '$lib/backend/projects';
 	import Resizer from '$lib/components/Resizer.svelte';
 	import { projectAiGenAutoBranchNamingEnabled } from '$lib/config/config';
 	import { projectAiGenEnabled } from '$lib/config/config';
@@ -35,6 +33,8 @@
 	import { onMount } from 'svelte';
 	import type { Persisted } from '$lib/persisted/persisted';
 	import type { Writable } from 'svelte/store';
+	import { Project } from '$lib/backend/projects';
+	import { User } from '$lib/stores/user';
 
 	export let isUnapplied = false;
 	export let isLaneCollapsed: Persisted<boolean>;
