+++ app/src/routes/+layout.ts
@@ -2,8 +2,8 @@ import { AIService } from '$lib/ai/service';
 import { initPostHog } from '$lib/analytics/posthog';
 import { initSentry } from '$lib/analytics/sentry';
 import { AuthService } from '$lib/backend/auth';
-import { HttpClient } from '$lib/backend/httpClient';
 import { GitConfigService } from '$lib/backend/gitConfigService';
+import { HttpClient } from '$lib/backend/httpClient';
 import { ProjectService } from '$lib/backend/projects';
 import { PromptService } from '$lib/backend/prompt';
 import { UpdaterService } from '$lib/backend/updater';
